Fix mask shift overflow

+++ src/ft_bits.c
@@ -6,7 +6,7 @@
 /*   By: clsaad <clsaad@student.42lyon.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/16 18:11:28 by clsaad            #+#    #+#             */
-/*   Updated: 2023/06/16 21:59:21 by clsaad           ###   ########.fr       */
+/*   Updated: 2023/06/26 17:30:43 by clsaad           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ static inline size_t	array_idx(size_t len, uint64_t seq)
 
 static inline uint64_t	bit_mask(uint64_t seq)
 {
-	return (0x01 << (seq % 64));
+	return ((uint64_t)0x01 << (seq % 64));
 }
 
 bool	is_bit_set(const uint64_t *bits_array, size_t len, uint64_t seq)
